crud completa, modifica a metri quadri

+++ database/migrations/2020_06_09_101655_create_info_places_table.php
@@ -18,7 +18,7 @@ class CreateInfoPlacesTable extends Migration
             $table->tinyInteger('rooms');
             $table->tinyInteger('beds');
             $table->tinyInteger('bathrooms');
-            $table->tinyInteger('m2')->nullable();
+            $table->smallInteger('m2')->nullable();
             $table->text('description');
             $table->timestamps();
         });
